scripts/requirements.txt
firebase-admin>=6.0.0
requests>=2.28.0
boto3>=1.26.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
